cli: Suggest to make new identity the default for server (#1109)
When a server key rotation is suspected, suggest to make the new
identity the default.

This usually happens during development using an ephemeral instance. If
one follows the instructions, the `default_identity` is not set for the
existing server, which makes the CLI generate a fresh identity every
time.

See also: #333