host: Unregister panicked modules from the host controller
Add a Fn object to `ModuleHost` which is invoked whenever a reducer call
on this host panics. When created from the `HostController`, the
function simply unregisters the module from the controller.

This should release both the module and database resources, such that
the database restarts on subsequent access. This is the intended
behavior to land #985.

Note that this patch does not address the disjoint lifecycles of modules
and databases, thus a panic when executing an SQL command is likely to
leave the database in a permanently inaccessible state.