Integrate snapshotting into core
This commit integrates snapshotting into `RelationalDB`.

In this initial version, we take a snapshot every 1 million txes written to the commitlog.

Part of this involves making the `committed_state` module `pub(crate)`
where previously it was private.
In order to preserve the previous visibility of its members,
anything that was previously `pub` is now `pub(super)`.

Notable TODOs, some or all of which may be left as future work:

- Mark snapshots newer than the `durable_tx_offset` as invalid when restoring.
- Write a test that exercises snapshot capture and restore.
  This is challenging because of the large `SNAPSHOT_INTERVAL`;
  unless/until we make that configurable, such a test would have to commit 1 million txes.
- Make `SNAPSHOT_INTERVAL` configurable.