Fix tests
Durable DBs have a new dependency on Tokio in order to run the snapshot worker.
This was causing tests to fail
because the std test harness does not install a Tokio runtime.

This commit fixes the test failures by installing a Tokio runtime
around every test which constructs a durable DB.