Add explicit parameter for table visibility
After this change, tables must be marked with `Public = true` if they're meant to be exposed to clients.

Note that the syntax is different from Rust because attributes are not macros with arbitrary syntax, but must be actual expressions.

For now boolean should work well enough, but we might revisit specific attribute structure in the future - e.g. if Rust changes to separate `#[spacetimedb(table)] #[spacetimedb(public)]`, then it should be easy enough to change C# to use 2 attributes as well.

Fixes #1347.