Ignore some mismatched seq allocations during update
When checking for schema compatibility before an update,
we erroneously considered sequence allocations to be an incompatible change.
This means that any module which contained at least one sequence/`autoinc` column
which had allocated more than 4096 values would be impossible to ever update.

With this change, we permit updates where the proposed sequence allocation
is less than the known sequence allocation,
and use the larger known sequence allocation as the value.
This makes it possible to update modules
which contain `autoinc` columns that have allocated more than 4096 values.