core: Fix module update
When updating a host, that host may not currently be running, in which
case it is launched. However, it may (or may not) be initialized
already. In either case, it is initialized before updating.

However, the host controller handled the not-running case without
considering the auto-initialization -- that is, it took the module the
host was launched with to perform the update, instead of the new module.

This meant that the schema was the same, so no actual update was
performed.

This patch fixes this by handling both cases uniformly, which always
instantiates a module from the supplied program bytes.

It also passes the correct program hash to `launch_module`, so as to fix
the diagnostic output.