Remove hardcoded types from C# codegen
We leverage the fact that all primitives in C# have a corresponding fully-qualified name, and declare a single namespace with all their BSATN implementations.

This makes type resolution in the codegen simpler - it doesn't need to hold an ever-growing switch of special types - and allows contributors to add new System implementations (e.g. for collections) without having to deal with Roslyn at all.