[C#] [NFC] Reuse attribute classes from runtime
Currently, we have to copy-paste and hardcode property names and types from the runtime in the codegen, which is pretty error-prone, and is getting harder as our API is going to rapidly expand.

In this PR I'm implementing a basic helper that allows to reuse original attribute classes from the runtime in the codegen by taking the syntax data and reconstructing the same attribute instance via reflection.

This ensures that we have all the same typechecked properties on both sides instead of having to extract them manually by name + explicit casts.