Revise `BTreeIndexBounds`: No non-`Copy` owned values; restrict filterable types
And add some comments to the macros which implement it.

For reasons that escape me, this breaks inference on integer literals:
they get defaulted too early to `i32`,
even when the indexed column is of a different type like `u32` or `i64`.
We could potentially work around this
by making `i32` a `FilterableValue` for any integer column type,
but that opens the door for some nasty truncation-related bugs.