Eliminate dependency on `im`; just mutate the client cache
This turns out to be effectively a prerequisite for client-side indices.
Storing indexes in a type-erased way
while making them `Clone` for storage in `im::HashMap`
is likely possible, but it's at least very challenging and complex.
Given that we no longer actually need the concurrent persistence offered by `im::HashMap`,
it seems easier to just revert to using `std::collections::HashMap`.