Fix Find() with struct-based rows
Before this, `Find()` would not return a nullable `Row?` when `Row` is a value type (a `struct`), leading to compilation errors at best and at a struct with all fields set to `0` or `null` at worst.

This issue was found while updating benchmarks-cs and it's not covered by other `modules/*` tests, which is why I bundled it with other fixes that make benchmarks-cs compile again in the same PR. I also made a change that ensures that `becnhmarks-cs` actually compiles in CI (before this, it could only be executed manually).

The fix for this particular issue is in `Codegen/Module.cs`, while other fixes are removing `IndexedAttribute` (which shouldn't be exposed to the users) and adding actual indexes to the benchmarks, updating table names, using new `.Find` / `.Filter` APIs and such.