Avoid div-by-zero in row count estimations for empty tables
Forward-port of an 0.11.1 hotfix for the BitCraft alpha.

What was happening was, `num_distinct_values` on an empty table returned 0,
which `index_row_est` then used as a divisor, leading to a panic.
It seems reasonably clear to me that
the correct estimate of distinct values in an empty index is 0, not bottom,
so adjusting `num_distinct_values` to return `Option<NonZeroU64>`,
and then making `index_row_est` return 0 in that case,
seems an obvious fix.