Remove problematic debug assertions
The definition of `FixedHeader` within `Page` previously contained
a `#[cfg(debug_assertions)]` member which was used for some extra assertions.
This was fine originally, but now that we write pages to disk in snapshots,
having the serialized format of the page vary depending on compilation profile
is problematic.

This commit removes said `#[cfg(debug_assertions)]` conditional member from `FixedHeader`,
so that the structure, and thus the on-disk format, of `Page` is consistent
between debug and release builds.

This commit also removes an unrelated but unused `#[cfg(debug_assertions)]` conditional
from the pointer map code. We never called this assertion, even in debug builds.
If we want to re-enable it in the future, we can restore it from git history;
for now, I would prefer not to have unused code committed.