Migrate C# client-side reducer enum to the new syntax
As per https://github.com/clockworklabs/SpacetimeDBPrivate/pull/1198.

This makes reducers use the same codegen structure as tables and other tuple types, but nests them inside a `Reducer` class as subclasses for pattern-matching, like we allow on tagged enums.

The reason we don't use a tagged enum here anymore is because we don't need BSATN sum type support for the Reducer enum, and it imposes unnecessary restrictions like max 256 elements.