Rewrite inaccurate comments about empty struct-like variants
I thought rustc was rejecting empty struct-like variants,
but it turns out that was the SATS `derive(Serialize, Deserialize)` macros,
which we don't use anyways for the `enum Reducer`.
Oh well; generating the unit structs is better,
I just avoided it because it was more work,
and now that work is done, so we might as well keep it.