Enforce .NET 8 SDK for new C# modules
This fixes an issue where users with .NET 9 SDK installed can't build C# modules even if .NET 8 SDK is installed as well.

Unfortunately, this covers only new projects created with `spacetime init` as that's the only place where we control the project directory. For everyone else, we print instructions with a link to the official configuration docs.