Updating unity-tests (#197)
## Description of Changes
This bumps the unity-test submodule
 
## API

No API Change

## Requires SpacetimeDB PRs

No special PRs, should work with latest master of SpacetimeDB.

## Testsuite
*If you would like to run the your SDK changes in this PR against a
specific SpacetimeDB branch, specify that here. This can be a branch
name or a link to a PR.*

SpacetimeDB branch name: master

## Testing
I opened up this commit of CircleGame in Unity on my own machine and ran
the tests. They seemed to work fine.