Implement the `identity` host call and `ctx.identity()` in Rust
This PR adds `ReducerContext::identity(&self) -> Identity`,
a method which reads the module identity. It's useful e.g.
so modules can assert that scheduled reducers are being called by the scheduler,
rather than by a client.
It's also a part of the API spec that fell by the wayside during 0.12.

This PR also adds a smoketest that said method does what it's supposed to,
and adds a call to it in the `rust-wasm-test` module that I used during manual testing.

Still TODO: implement the same method in C# modules.