How-to: Incremental Migrations
This commit adds a how-to guide for defining "incremental migrations,"
a strategy for updating the schema of a database while maintaining compatibility
with outdated clients and without requiring a manual migration.

The code is not on GitHub yet,
as I'd like review on my choice of example before pushing the repository.
As such, the links to the code at the bottom of the new document are broken.