`Table::num_rows` and `Table::bytes_used_by_rows`
We intend to bill based on these predictable metrics,
rather than the somewhat-unpredictable actual heap memory usage of the system.
As such, we need a way to compute them (duh).
This commit adds `Table` methods for computing the number of resident rows,
and the number of bytes stored by those rows.