Use `println`/`eprintln` in the quickstart-chat client, not `log`
Likely by mistake, #1929 rewrote the `quickstart-chat` client
to use `log::info!` and `log::error!` for output.
In addition to needlessly adding complexity to a quickstart project,
this was broken because that program didn't install a logger,
so it produced no output at all.

This commit reverts those changes, returning to `println`/`eprintln` output
in the quickstart-chat client.
We could use `log::` macros and install a logger,
but then we'd have to explain installing a logger in the quickstart guide.

This commit also makes me a codeowner of the `quickstart-chat` project,
both the module and the client.
I think extra scrutiny on these files is important
because they're always supposed to exactly match the code presented in the tutorials,
which prior to this commit they clearly did not.