Nix `MapRowTypeToTable`; make table handles nameable
In a noble but misguided effort to make table handles not nameable,
previous versions of our `bindings-macro` `#[table]` attribute
hid the table struct away in a `const _: () = {}` block,
and used a hidden trait `MapRowTypeToTable` to name it
where necessary in the macro output.

This imposed the requirement that each row type corresponded to exactly one table.
This is not the only part of the macros that imposes that requirement,
so further commits will follow.