Make `DESCRIBERS` hold closures, not function pointers
Prior to this commit, `DESCRIBERS` was a vec of function pointers.
This meant that each `DESCRIBER` needed to be its own function,
rather than a closure over the thing being described.

With this commit, `DESCRIBERS` is amended to hold boxed closures.
This allows `register_row_level_security` to operate without a type parameter.