Don't create indexes during bootstrapping; wait until after replay
This PR fixes a bug where replaying databases without a snapshot of TX 0,
i.e. using the entire commitlog,
would fail when replaying deletes on `st_client`.

This is because `st_client` was in the unique (*) position of:
- Being a system table,
- with a unique index,
- which is mutated after bootstrapping.

(*) Other system tables meet all three criteria in edge cases
with schema-changing publishes, but only `st_client` meets them in everyday testing.

In this case, when replaying the entire commitlog,
we create the indexes on the system tables during bootstrapping before replay.
(I theorize, but have not checked,
that when restoring from a snapshot we do not create these indexes so eagerly,
instead waiting for after replaying the commitlog suffix to create them.)

However, we cannot update or use indexes during bootstrapping,
at least currently, because doing so causes unique constraint violations.
Prior to this commit, we had a bunch of complicated code
which avoided updating indexes during replay.

PR #2092, aka commit f667f65b159f0ffbd00a3bdb24afee79c0084070,
introduced a case where we would use an index during replay:
when replaying a delete for a table which had an index,
and therefore had no pointer map.
Prior to that commit, we would never use an index during replay, even if it existed.

This all came together to mean that `st_client` had an empty index, and no pointer map,
and so calls to `replay_delete_by_rel` on that table would always fail.
I theorize that no one noticed this sooner
because it did not manifest when restoring from a snapshot,
and there is always a snapshot of TX 0 unless you manually delete it,
as Kim did during his testing.

The fix is simple: we do not construct indexes on the system tables
during bootstrapping before replaying the commitlog,
instead waiting until after replay like we do for user-defined indexes.
This allows us to rip out all the aforementioned fiddly code
for dodging indexes during replay,
since now there will not be any indexes and therefore we will never update them.